# Conflicts:
#	atlas-core/build.gradle

+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tasks/transform/dex/AtlasMainDexMerger.java
@@ -16,6 +16,7 @@ import com.android.builder.utils.FileCache;
 import com.android.ide.common.process.ProcessException;
 import com.android.ide.common.process.ProcessOutput;
 import com.android.ide.common.process.ProcessOutputHandler;
+import com.android.prefs.AndroidLocation;
 import com.android.utils.FileUtils;
 import com.taobao.android.builder.tools.ReflectUtils;
 import org.gradle.api.Project;
@@ -50,12 +51,15 @@ public class AtlasMainDexMerger extends AtlasDexMerger {
             public void handleMissActionResult(File outputDir, File in) throws IOException {
                 File[] dexs = outputDir.listFiles((dir, name) -> name.endsWith(".dex"));
                 if (dexs != null && dexs.length > 1) {
-                    for (File file : dexs) {
-                        org.apache.commons.io.FileUtils.copyFileToDirectory(file, in.getParentFile());
-                    }
+                    org.apache.commons.io.FileUtils.copyDirectory(outputDir,in);
+//                    for (File file : dexs) {
+//                        org.apache.commons.io.FileUtils.copyFileToDirectory(file, in.getParentFile());
+//                    }
 
                 } else if (dexs.length == 1) {
-
+                    if (!in.getParentFile().exists()){
+                        in.getParentFile().mkdirs();
+                    }
                     Files.copy(dexs[0].toPath(), in.toPath());
                 } else {
                     throw new IOException("no dex file merge out");
@@ -66,18 +70,18 @@ public class AtlasMainDexMerger extends AtlasDexMerger {
             public void handleQueryResult(FileCache.QueryResult result, File outputDir, String bundleName) throws IOException {
 
                 if (result.getQueryEvent().equals(FileCache.QueryEvent.HIT)) {
-                    if (result.getCachedFile().exists()) {
+                    if (result.getCachedFile().exists() && result.getCachedFile().isDirectory()) {
                         logger.info("hit maindex dexmerge cache ->" + result.getCachedFile().getAbsolutePath());
-                        org.apache.commons.io.FileUtils.copyFile(result.getCachedFile(), new File(outputDir, CLASSES_DEX));
-                    } else if (!result.getCachedFile().exists()) {
-                        File[] dexs = result.getCachedFile().getParentFile().listFiles((dir, name) -> name.endsWith(".dex"));
-                        if (dexs != null && dexs.length > 0) {
-                            for (File dex : dexs) {
-                                logger.info("hit maindex dexmerge cache ->" + dex.getAbsolutePath());
-                                org.apache.commons.io.FileUtils.copyFileToDirectory(dex, outputDir);
-                            }
+                        org.apache.commons.io.FileUtils.copyDirectory(result.getCachedFile(), outputDir);
+                    } else if (result.getCachedFile().exists() && result.getCachedFile().isFile()) {
+//                        File[] dexs = result.getCachedFile().getParentFile().listFiles((dir, name) -> name.endsWith(".dex"));
+//                        if (dexs != null && dexs.length > 0) {
+//                            for (File dex : dexs) {
+                                logger.info("hit maindex dexmerge cache ->" + result.getCachedFile().getAbsolutePath());
+                                org.apache.commons.io.FileUtils.copyFile(result.getCachedFile(), new File(outputDir,CLASSES_DEX));
+//                            }
 
-                        }
+//                        }
                     }
                 } else {
                     logger.info("miss maindex dexmerge cache -> null");
@@ -130,6 +134,13 @@ public class AtlasMainDexMerger extends AtlasDexMerger {
                         mainDexFiles, dexingType, dexMerger, mainDexListFile == null ? null : mainDexListFile.getSingleFile(), minSdkVersion, isDebuggable, "maindex", ID);
         ProcessOutput output = outputHandler.createOutput();
         FileCache fileCache = BuildCacheUtils.createBuildCacheIfEnabled(variantOutputContext.getVariantContext().getProject(), variantOutputContext.getScope().getGlobalScope().getProjectOptions());
+        if (fileCache == null){
+            try {
+                fileCache = FileCache.getInstanceWithMultiProcessLocking(new File(AndroidLocation.getFolder(),"atlas-buildCache"));
+            } catch (AndroidLocation.AndroidLocationException e) {
+                e.printStackTrace();
+            }
+        }
         try {
             FileCache.QueryResult result = fileCache.createFileInCacheIfAbsent(
                     buildCacheInputs,
@@ -168,13 +179,7 @@ public class AtlasMainDexMerger extends AtlasDexMerger {
     }
 
 
-    @NonNull
-    private File getDexOutputLocation(
-            @NonNull TransformOutputProvider outputProvider,
-            @NonNull String name,
-            @NonNull Set<? super QualifiedContent.Scope> scopes) {
-        return outputProvider.getContentLocation(name, TransformManager.CONTENT_DEX, scopes, Format.DIRECTORY);
-    }
+
 
 
 }
